Created Validation and Theme applied
User input must meet a certain criteria in order to successfully add a contact. Also changed the colors of the new contact form to better fit the bread theme. There still is more to be done when it comes to ux/accessibility. Search, Update, and Delete still need to be done.